codes/RLProject-main/requirements.txt
wheel
gym-retro
opencv-python
tensorflow==2.5.0
scipy
numpy
scikit-image
